# ssoapi/nodeapi-swagger.yaml
swagger: '2.0'
info:
  title: 'SSO API'
  version: '1.0'
basePath: /
schemes:
  - http
paths:
  /oauth/token:
    post:
      consumes:
        - application/x-www-form-urlencoded
      parameters:
        - description: Grant Type
          in: formData
          name: grant_type
          type: string
          enum: [password]
          required: true
        - description: Client Id
          in: formData
          name: client_id
          type: string
          enum: [baf50a1c-1a7d-11ea-978f-2e728ce88125]
          required: true
        - description: Email
          in: formData
          name: username
          required: true
          type: string
        - description: Password
          in: formData
          name: password
          required: true
          type: string
      responses:
        '200':
          description: |
            <pre>
              {
                "error": null,
                "status": 1,
                "response": {
                  "accessToken": "String",
                  "accessTokenExpiresAt": "String",
                  "refreshToken": "String",
                  "refreshTokenExpiresAt": "String",
                  "scope": "String",
                  "client": {
                    "id": "Integer"
                  },
                  "user": {
                    "id": "Integer"
                  }
                }
              }              
            </pre>
        '500':
          description: |
            <pre>
              {
                "error": "Invalid User",
                "status": 0,
                "response": null
              }
            </pre>
      tags:
        - Auth
      description:
      operationId: 
      summary: authenticate an user
  /oauth/me:
    get:
      consumes:
        - application/x-www-form-urlencoded
      parameters:
        - description: Bearer Token
          in: header
          name: Authorization
          type: string
          required: true
      responses:
        '200':
          description: |
            <pre>
              {
                "error": null,
                "status": 1,
                "response": {
                  "accessToken": "String",
                  "accessTokenExpiresAt": "String",
                  "scope": "String",
                  "client": {
                    "id": 1,
                    "redirect_uris": null,
                    "grants": [
                      "password"
                    ],
                    "user_id": null
                  },
                  "user": {
                    "id": "Integer",
                    "tenant_id": "Integer",
                    "username": "String",
                    "name": "String",
                    "phone": "String",
                    "scope": "String"
                  }
                }
              }              
            </pre>
        '401':
          description: |
            <pre>
              {
                "error": "Unauthorized",
                "status": 0,
                "response": null
              }
            </pre>
      tags:
        - Auth
      description:
      operationId: 
      summary: get authenticated user
...